+++ CHANGELOG.md
@@ -33,6 +33,7 @@ This build is not intended for submission, but rather for the developers to keep
 - Investments bought are assigned to the player
 - Game loop bug fixed when player lands on their own property
 - There is an effort in fixing "drunk rent crashes".
+- The game would now check for *some* winning conditions.
     
 ## Known issues
 
@@ -43,6 +44,7 @@ This build is not intended for submission, but rather for the developers to keep
 - changes to button class may cause issues (might need to create an inheritance class for seperate buttons)
 - investment fields are not changing names on purchase even though ownership changes
   ~~ - infinite glitch occurs when player lands on their own investment. ~~ Fixed
+- Some winning conditions haven't been implemented yet.
 
 ## Potential improvements: